# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/api/v1/api-keys": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns all API keys",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "List API keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/mdb.ApiKey"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Create a new universal API key. Usable for both gateway flows (epay/ezpay). PID auto-incremented; secret returned once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "description": "API key payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.CreateApiKeyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}": {
            "delete": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Soft-delete an API key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Delete API key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Patch API key name / ip_whitelist / notify_url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Update API key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UpdateApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/rotate-secret": {
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Generate a new secret_key. Returned only once.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Rotate API key secret",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.RotateSecretResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/secret": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns the secret_key for an API key (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "View API key secret",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/stats": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns usage stats (call_count, last_used_at) for an API key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Get API key stats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.ApiKeyStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/status": {
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Toggle enable/disable for an API key (1=enabled, 2=disabled)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Change API key status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.ChangeApiKeyStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.ApiKeyStatsResponse": {
            "type": "object",
            "properties": {
                "call_count": {
                    "type": "integer",
                    "example": 342
                },
                "last_used_at": {
                    "type": "string",
                    "example": "2026-04-14T08:30:00+00:00"
                }
            }
        },
...
        "admin.ChangeApiKeyStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态 1=启用 2=禁用",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                }
            }
        },
...
        "admin.CreateApiKeyRequest": {
            "type": "object",
            "properties": {
                "ip_whitelist": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": "My API Key"
                },
                "notify_url": {
                    "type": "string",
                    "example": "https://example.com/notify"
                }
            }
        },
        "admin.CreateApiKeyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My API Key"
                },
                "pid": {
                    "type": "string",
                    "example": "1003"
                },
                "secret_key": {
                    "description": "SecretKey is returned ONCE on creation. After that, fetch via\nGET /api-keys/:id/secret or rotate to generate a new one.",
                    "type": "string",
                    "example": "secret123abc456"
                }
            }
        },
...
        "admin.RotateSecretResponse": {
            "type": "object",
            "properties": {
                "secret_key": {
                    "type": "string",
                    "example": "newSecret123abc456"
                }
            }
        },
...
        "admin.UpdateApiKeyRequest": {
            "type": "object",
            "properties": {
                "ip_whitelist": {
                    "type": "string",
                    "example": "10.0.0.1,192.168.0.0/24"
                },
                "name": {
                    "type": "string",
                    "example": "Updated Key Name"
                },
                "notify_url": {
                    "type": "string",
                    "example": "https://example.com/notify"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "mdb.ApiKey": {
            "type": "object",
            "properties": {
                "call_count": {
                    "type": "integer",
                    "example": 342
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-16 12:00:00"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "ip_whitelist": {
                    "type": "string",
                    "example": "192.168.1.0/24,10.0.0.1"
                },
                "last_used_at": {
                    "type": "string",
                    "example": "2026-04-16 12:00:00"
                },
                "name": {
                    "type": "string",
                    "example": "My API Key"
                },
                "notify_url": {
                    "type": "string",
                    "example": "https://example.com/notify"
                },
                "pid": {
                    "type": "string",
                    "example": "1000"
                },
                "status": {
                    "description": "状态 1=启用 2=禁用",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-16 12:00:00"
                }
            }
        },
...
        "response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    }
}